chore: move autoEnv to platform.info.
