+++ packages/sdk/react-native/src/platform/createAutoEnv.ts
