+++ packages/sdk/react-native/src/platform/index.ts
@@ -22,6 +22,7 @@ import { name, version } from '../../package.json';
 import { btoa, uuidv4 } from '../polyfills';
 import RNEventSource from '../react-native-sse';
 import AsyncStorage from './ConditionalAsyncStorage';
+import createAutoEnv from './createAutoEnv';
 
 class PlatformRequests implements Requests {
   createEventSource(url: string, eventSourceInitDict: EventSourceInitDict): EventSource {
@@ -46,6 +47,7 @@ class PlatformInfo implements Info {
   platformData(): PlatformData {
     return {
       name: 'React Native',
+      autoEnv: createAutoEnv(),
     };
   }
